Extract toolkit-indepenent code into `SAM.Core`
