+++ SAM.Picker/SAM.Picker.csproj
@@ -53,5 +53,6 @@
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\SAM.API\SAM.API.csproj" />
+    <ProjectReference Include="..\SAM.Core\SAM.Core.csproj" />
   </ItemGroup>
 </Project>
